misc/R2A/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
beartype==0.14.1
better-abc==0.0.3
certifi==2025.8.3
charset-normalizer==3.4.2
circuitsvis==1.43.3
click==8.2.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.16
decorator==5.2.1
dill==0.3.8
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fancy-einsum==0.0.3
filelock==3.13.1
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jaxtyping==0.3.2
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
narwhals==2.5.0
nest_asyncio==1.6.0
networkx==3.3
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.0
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
pip==25.1
platformdirs==4.3.8
plotly==6.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scipy==1.16.2
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.32.0
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
stack_data==0.6.3
sympy==1.13.3
tokenizers==0.21.2
torch==2.4.1+cu121
torchvision==0.19.1+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
transformers-stream-generator==0.0.5
triton==3.0.0
typeguard==4.4.4
typing_extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wadler_lindig==0.1.7
wandb==0.21.0
wcwidth==0.2.14
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
